# REST-API/cat/static/swagger.json
openapi: 3.0.3
info:
  title: CAT
  description: Case Assignment Tool
  termsOfService: http://swagger.io/terms/
  contact:
    email: himanshu.surendra@gmail.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
servers:
  - url: /
...
paths:
...
  /api/product/{productname}:
    get:
      tags:
        - product
      summary: Get product details by productname
      description: ''
      operationId: getProductByName
      parameters:
        - name: productname
          in: path
          description: The product details that needs to be fetched.
          required: true
          schema:
            type: string
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
            application/xml:
              schema:
                $ref: '#/components/schemas/Product'
        '404':
          description: Object not found
    post:
      tags:
        - product
      summary: Update product
      description: ''
      operationId: updateProduct
      parameters:
        - name: productname
          in: path
          description: The product name that needs to be edited
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductModel'
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
            application/xml:
              schema:
                $ref: '#/components/schemas/Product'
        '400':
          description: Invalid Input
        '404':
          description: Object not found
        '401':
          description: Unauthorized Access
    delete:
      tags:
        - product
      summary: Delete product
      description: ''
      operationId: deleteProduct
      parameters:
        - name: productname
          in: path
          description: The product name that needs to be deleted
          required: true
          schema:
            type: string
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
            application/xml:
              schema:
                $ref: '#/components/schemas/Product'
        '400':
          description: Invalid Input
        '404':
          description: Object not found
        '401':
          description: Unauthorized Access
  /api/users-of-product/{productname}:
    get:
      tags:
        - product
      summary: Get all users list supporting product productname
      description: ''
      operationId: getUsersOfProduct
      parameters:
        - name: productname
          in: path
          description: The productname who users list needs to be fetched
          required: true
          schema:
            type: string
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
            application/xml:
              schema:
                $ref: '#/components/schemas/User'
  /api/product:
    post:
      tags:
        - product
      summary: Add a new product
      description: ''
      operationId: addProduct
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddProduct'
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
            application/xml:
              schema:
                $ref: '#/components/schemas/Product'
        '400':
          description: Invalid Input
  /api/product/schedule-sf-integration/{productname}:
    post:
      tags:
        - product
      summary: Schedule SF case polling for product productname
      description: ''
      operationId: scheduleSfIntegration
      parameters:
        - name: productname
          in: path
          description: The productname that needs to be scheduled for sf polling
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScheduleSFIntegration'
      responses:
        '200':
          description: successful operation
        '400':
          description: Invalid Input
  /api/product/delete-job/{jobid}:
    delete:
      tags:
        - product
      summary: Delete a scheduled job specific to a product
      description: ''
      operationId: productdeleteJob
      parameters:
        - name: jobid
          in: path
          description: The jobid of job that needs to be deleted
          required: true
          schema:
            type: string
      responses:
        '200':
          description: successful operation
        '404':
          description: Not Found
...
  /api/assign-case:
    post:
      tags:
        - case
      summary: Assign case
      description: ''
      operationId: assignACase
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssignCaseReq'
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Case'
            application/xml:
              schema:
                $ref: '#/components/schemas/Case'
        '400':
          description: Invalid Input
        '404':
          description: Object not found
  /api/unassign-case/{id}:
    post:
      tags:
        - case
      summary: Unassign case
      description: ''
      operationId: unassignCase
      parameters:
        - name: id
          in: path
          description: The case id that needs to be unassigned
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Comments'
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Case'
            application/xml:
              schema:
                $ref: '#/components/schemas/Case'
        '400':
          description: Invalid Input
        '404':
          description: Object not found
  /api/cases:
    get:
      tags:
        - case
      summary: Get list of all cases and their details
      description: ''
      operationId: getCases
      parameters:
        - in: query
          name: page
          schema:
            type: integer
          description: page number to return data from
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Case'
            application/xml:
              schema:
                $ref: '#/components/schemas/Case'
  /api/case/{id}:
    get:
      tags:
        - case
      summary: Get case by case id
      description: ''
      operationId: getCaseDetails
      parameters:
        - name: id
          in: path
          description: The case id whose details needs to be fetched.
          required: true
          schema:
            type: string
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Case'
            application/xml:
              schema:
                $ref: '#/components/schemas/Case'
        '404':
          description: Case not found
  /api/cases-of-product/{productname}:
    get:
      tags:
        - case
      summary: Get all cases of product productname
      description: ''
      operationId: getCasesOfProduct
      parameters:
        - name: productname
          in: path
          description: productname whose list of cases needs to be fetched
          required: true
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
          description: page number to return data from
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Case'
            application/xml:
              schema:
                $ref: '#/components/schemas/Case'
        '404':
          description: Case not found
  /api/cases-of-team/{teamname}:
    get:
      tags:
        - case
      summary: Get cases of all users belonging to team teamname.
      description: ''
      operationId: getCasesOfTeam
      parameters:
        - name: teamname
          in: path
          description: teamname whose list of cases needs to be fetched
          required: true
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
          description: page number to return data from
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Case'
            application/xml:
              schema:
                $ref: '#/components/schemas/Case'
        '404':
          description: Case not found
  /api/cases-assigned-by-user/{username}:
    get:
      tags:
        - case
      summary: Get all cases assigned by user username
      description: ''
      operationId: getCasesAssignedByUser
      parameters:
        - name: username
          in: path
          description: username whose list of assigned_by cases needs to be fetched
          required: true
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
          description: page number to return data from
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Case'
            application/xml:
              schema:
                $ref: '#/components/schemas/Case'
        '404':
          description: Case not found
  /api/cases-of-user/{username}:
    get:
      tags:
        - case
      summary: Get all cases assigned to user username
      description: ''
      operationId: getCasesOfUser
      parameters:
        - name: username
          in: path
          description: username whose list of assigned cases needs to be fetched
          required: true
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
          description: page number to return data from
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Case'
            application/xml:
              schema:
                $ref: '#/components/schemas/Case'
        '404':
          description: Case not found
  /api/case-count-of-all-users:
    post:
      tags:
        - case
      summary: Get case count of users over a period listed by interval. Can be further filtered by product and/or team.
      description: ''
      operationId: getCaseCountOfAllUsers
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CaseCountReq'
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CaseCount'
            application/xml:
              schema:
                $ref: '#/components/schemas/CaseCount'
        '400':
          description: Invalid Input
        '404':
          description: Object not found
components:
  schemas:
    AssignCaseReq:
      required:
        - case_id
        - product
      type: object
      properties:
        case_id:
          type: string
          example: 2023-0910-029384
        product:
          type: string
          example: Product1
        priority:
          type: string
          example: p2
        comments:
          type: string
          example: this is an example
        mode:
          type: string
          enum:
            - auto
            - manual
            - auto-reassign
          description: defaults to auto. if manual, specify user to assign the case to. If auto-reassign, the case gets auto-reassigned if already cat'ed or gets assigned if not already cat'ed
        user:
          type: string
          description: username to assign case to if mode is manual
        check_in_shift:
          type: boolean
          description: whether to check if user is in shift or not when assiging a case
        sf_account_name:
          type: string
          description: account name in salesforce
          example: xyz
        sf_email_name:
          type: string
          description: j2 email template name
          example: email1
        delayed_assignment:
          type: string
          description: time to consider to check which users are in shift when assigning case
          example: 16:00:00
    UserModel:
      type: object
      properties:
        email:
          type: string
          example: abc@efg.com
        teamname:
          type: string
          example: Team1
        active:
          type: boolean
          description: default is 1
        admin:
          type: boolean
          description: whether the user is admin or not. Default is false
        team_email_notifications:
          type: boolean
          description: whether the user recives team activity notifications. Default is false
        monitor_case_updates:
          type: boolean
          description: Monitor users cases for external updates when user is ooo or on weekends. Default is false
        shift_start:
          type: string
          example: '09:00:00'
        shift_end:
          type: string
          example: '18:00:00'
        first_name:
          type: string
          example: 'Tom'
        last_name:
          type: string
          example: 'Kumar'
        timezone:
          type: string
          example: US/Pacific
          description: python lib pytz format
    User:
      allOf:
        - type: object
        - $ref: '#/components/schemas/UserModel'
      properties:
        username:
          type: string
          example: u1
        user_since:
          type: string
          description: default is current timestamp when adding a new user.
          example: Fri, 06 Oct 2023 01:58:56 GMT
        products:
          type: string
          example: Product1
        full_name:
          type: string
          example: Bob Marley
        last_login:
          type: string
          description: default is user creation timestamp. keeps track of users last login.
          example: Sat, 07 Oct 2023 12:45:56 GMT
        in_shift:
          type: boolean
          description: whether the user is in shift or not when the call is made
          example: true
...
    Case:
      type: object
      properties:
        case_id:
          type: string
          example: 2023-0910-02938
        assigned_by:
          type: string
          example: Tom
        assigned_to:
          type: string
          example: Cody
        product:
          type: string
          example: Product1
        time:
          type: string
          description: Default is current timestamp when assigning a case
          example: Mon, 02 Oct 2023 16:56:10 GMT
        mode:
          type: string
          enum:
            - auto
            - manual
        comments:
          type: string
          example: Testing
        priority:
          type: string
          example: p2
        sf_account_name:
          type: string
          example: xyz
        assignment_history:
          type: string
          example: ['ba','cd','ef']
          description: string which is a list of usernames the case has been cat'ed to so far
...
    ProductModel:
      type: object
      properties:
        strategy:
          type: string
          enum:
            - s1
            - s2
          description: strategy for distributing cases
        max_days:
          type: string
          example: 3
          description: maximum number of days to look back when trying to find the user to assign a case to. [look up is one day at a time]
        max_days_month:
          type: string
          example: 300
          description: maximum number of days to look back when trying to find the user to assign a case to. [look up is one month at a time]
        case_regex:
          type: string
          example: ^[0-9]{4}-[0-9]{4}-[0-9]{6}$
        quota_over_days:
          type: integer
          description: when quota is factored in, this variable is used to find how may days to look back. default is 1
          example: 1
        sf_api:
          type: string
          example: Salesforce API for fetching list of cases
        sf_job_cron:
          type: string
          example: "* 6-17 * * 1-5"
        sf_job_timezone:
          type: string
          example: US/Pacific
        sf_product_series:
          type: string
          example: Session Smart Router
        sf_platform:
          type: string
          example: Session Smart Router
        sf_job_query_interval:
          type: integer
          description: how often to query sf api in mins
          example: 1
        sf_enabled:
          type: boolean
          description: Integrate with Sales
          example: true
        sf_init_email_name:
          type: string
          description: (optional) the email template name to use when sending initial response on a case beloging to an sf integrated product
          example: email1
    Product:
      type: object
      allOf:
        - $ref: '#/components/schemas/ProductModel'
        - type: object
      properties:
        users:
          type: string
          example: user1
        productname:
          type: string
          example: Product1
    AddProduct:
      type: object
      allOf:
        - $ref: '#/components/schemas/ProductModel'
        - type: object
      required:
        - productname
      properties:
        productname:
          type: string
          example: Product1
...
    CaseCountReq:
      type: object
      required:
        - period
        - interval
      properties:
        period:
          type: integer
          format: int64
          example: 30
          description: in days
        interval:
          type: integer
          format: int64
          example: 1
          description: in days
        productname:
          type: string
          example: Product1
        teamname:
          type: string
          example: Team1
    CaseCount:
      type: object
      additionalProperties:
        type: integer
...
    ScheduleSFIntegration:
      type: object
      required:
        - datetime
        - timezone
        - sf_enabled
      properties:
        datetime:
          type: string
          example: 2023/10/19 13:12
        timezone:
          type: string
          example: US/Pacific
        sf_enabled:
          type: boolean
          example: true
        holiday_list:
          type: string
          example: {"New Years": "2024-01-01", "MLK": "2024-01-15"}
          description: json formated list of holidays
    Comments:
      type: object
      properties:
        comments:
          type: string
          example: testing
...